Rename defaultFacetDistribution to initialFacetDistribution
