+++ packages/instant-meilisearch/src/cache/init-facets-distribution.ts
